# openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/pharmacies": {"get": {"summary": "Read Pharmacies", "operationId": "read_pharmacies_pharmacies_get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "exclusiveMinimum": 0, "default": 10, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/PharmacyWithHours"}, "title": "Response Read Pharmacies Pharmacies Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/pharmacies/{pharmacy_id}/masks": {"get": {"summary": "Read Masks By Pharmacy", "operationId": "read_masks_by_pharmacy_pharmacies__pharmacy_id__masks_get", "parameters": [{"name": "pharmacy_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Pharmacy Id"}}, {"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "exclusiveMinimum": 0, "default": 10, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Transaction"}, "title": "Response Read Masks By Pharmacy Pharmacies  Pharmacy Id  Masks Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/pharmacies_by_count_and_range": {"get": {"summary": "Read Masks By Count Abd Range", "operationId": "read_masks_by_count_abd_range_pharmacies_by_count_and_range_get", "parameters": [{"name": "min_count", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Min Count"}}, {"name": "max_count", "in": "query", "required": true, "schema": {"type": "integer", "exclusiveMinimum": 0, "title": "Max Count"}}, {"name": "min_price", "in": "query", "required": true, "schema": {"type": "number", "exclusiveMinimum": 0.0, "title": "Min Price"}}, {"name": "max_price", "in": "query", "required": true, "schema": {"type": "number", "exclusiveMinimum": 0.0, "title": "Max Price"}}, {"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "exclusiveMinimum": 0, "default": 10, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/PharmacyWithCount"}, "title": "Response Read Masks By Count Abd Range Pharmacies By Count And Range Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/top_user_amount": {"get": {"summary": "Read Top User Amount", "operationId": "read_top_user_amount_top_user_amount_get", "parameters": [{"name": "start_date", "in": "query", "required": true, "schema": {"type": "string", "format": "date-time", "title": "Start Date"}, "example": "2021-01-01"}, {"name": "end_date", "in": "query", "required": true, "schema": {"type": "string", "format": "date-time", "title": "End Date"}, "example": "2021-01-31"}, {"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "exclusiveMinimum": 0, "default": 10, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/UserTopCount"}, "title": "Response Read Top User Amount Top User Amount Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/transactions/summary": {"get": {"summary": "Read Transactions Summary", "operationId": "read_transactions_summary_transactions_summary_get", "parameters": [{"name": "start_date", "in": "query", "required": true, "schema": {"type": "string", "format": "date-time", "title": "Start Date"}, "example": "2021-01-01"}, {"name": "end_date", "in": "query", "required": true, "schema": {"type": "string", "format": "date-time", "title": "End Date"}, "example": "2021-01-31"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionSummary"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/search": {"get": {"summary": "Search Pharmacies And Masks", "operationId": "search_pharmacies_and_masks_search_get", "parameters": [{"name": "search_term", "in": "query", "required": true, "schema": {"type": "string", "title": "Search Term"}}, {"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "exclusiveMinimum": 0, "default": 10, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/PharmacyOrMask"}, "title": "Response Search Pharmacies And Masks Search Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/purchase": {"post": {"summary": "Purchase Masks", "operationId": "purchase_masks_purchase_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PurchaseRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionId"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "MaskWithPrice": {"properties": {"name": {"type": "string", "title": "Name"}, "id": {"type": "integer", "title": "Id"}, "price": {"type": "number", "title": "Price"}}, "type": "object", "required": ["name", "id", "price"], "title": "MaskWithPrice"}, "PharmacyHour": {"properties": {"day_of_week": {"type": "string", "title": "Day Of Week"}, "open_time": {"type": "string", "format": "time", "title": "Open Time"}, "close_time": {"type": "string", "format": "time", "title": "Close Time"}, "id": {"type": "integer", "title": "Id"}, "pharmacy_id": {"type": "integer", "title": "Pharmacy Id"}}, "type": "object", "required": ["day_of_week", "open_time", "close_time", "id", "pharmacy_id"], "title": "PharmacyHour"}, "PharmacyOrMask": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "type": {"type": "string", "title": "Type"}}, "type": "object", "required": ["id", "name", "type"], "title": "PharmacyOrMask"}, "PharmacyWithCount": {"properties": {"name": {"type": "string", "title": "Name"}, "cash_balance": {"type": "number", "title": "Cash Balance"}, "id": {"type": "integer", "title": "Id"}, "mask_count": {"type": "integer", "title": "Mask Count"}}, "type": "object", "required": ["name", "cash_balance", "id", "mask_count"], "title": "PharmacyWithCount"}, "PharmacyWithHours": {"properties": {"name": {"type": "string", "title": "Name"}, "cash_balance": {"type": "number", "title": "Cash Balance"}, "id": {"type": "integer", "title": "Id"}, "opening_hours": {"items": {"$ref": "#/components/schemas/PharmacyHour"}, "type": "array", "title": "Opening Hours", "default": []}}, "type": "object", "required": ["name", "cash_balance", "id"], "title": "PharmacyWithHours"}, "PurchaseRequest": {"properties": {"user_id": {"type": "integer", "title": "User Id"}, "mask_id": {"type": "integer", "title": "Mask Id"}, "pharmacy_id": {"type": "integer", "title": "Pharmacy Id"}}, "type": "object", "required": ["user_id", "mask_id", "pharmacy_id"], "title": "PurchaseRequest"}, "Transaction": {"properties": {"transaction_amount": {"type": "number", "title": "Transaction Amount"}, "date": {"type": "string", "format": "date-time", "title": "Date"}, "id": {"type": "integer", "title": "Id"}, "user_id": {"type": "integer", "title": "User Id"}, "pharmacy_id": {"type": "integer", "title": "Pharmacy Id"}, "mask_id": {"type": "integer", "title": "Mask Id"}, "mask": {"$ref": "#/components/schemas/MaskWithPrice"}}, "type": "object", "required": ["transaction_amount", "date", "id", "user_id", "pharmacy_id", "mask_id", "mask"], "title": "Transaction"}, "TransactionId": {"properties": {"transaction_id": {"type": "integer", "title": "Transaction Id"}}, "type": "object", "required": ["transaction_id"], "title": "TransactionId"}, "TransactionSummary": {"properties": {"total_amount": {"type": "integer", "minimum": 0.0, "title": "Total Amount", "default": 0}, "total_value": {"type": "number", "title": "Total Value"}}, "type": "object", "required": ["total_value"], "title": "TransactionSummary"}, "UserTopCount": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "total_amount": {"type": "integer", "title": "Total Amount"}}, "type": "object", "required": ["id", "name", "total_amount"], "title": "UserTopCount"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}